refactor: Refactor ApplicationAdapter to use TypeScript
Convert ApplicationAdapter from JavaScript to TypeScript to take advantage of
static typing and improve code maintainability. Update headers to use
consistent camelCase naming convention and explicitly declare service
dependencies. Also, update class property declarations to TypeScript syntax.
